Used to organize screenshots
Switch screenshots are saved in this format: [time]-[game id].jpg

You can add your own names by taking a screenshot, and looking at the screenshot file name. Then, you can add it to this list! Please enter it in the correct format (Needs to be written within the list)